deprecate stateful and rewrite in ts
